pyproject.toml
[project]
name = "spacenit"
version = "0.1.0"
description = "Tools for building geospatial foundation models"
authors = [{ name = "SpaceNit Team" }]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11, <3.14" # torch 2.7.x compatibility
dependencies = [
    # Minimal inference dependencies
    "einops>=0.7.0",
    "huggingface_hub",
    "numpy>=1.26.4",
    "torch>=2.7,<2.8",
    "universal-pathlib>=0.2.5",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["spacenit*"]

[tool.setuptools.package-data]
"spacenit" = ["py.typed"]
"spacenit.ingestion.norm_configs" = ["*.json"]

[tool.ruff]
fix = true

[tool.ruff.lint]
extend-select = ["D", "I", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.uv.sources]
claymodel = { git = "https://github.com/Clay-foundation/model.git" }
geobench = { git = "https://github.com/ServiceNow/geo-bench.git" }
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
training = [
    "ai2-olmo-core==2.3.0",
    # Training + data deps
    "albumentations>=1.4.10",
    "cartopy>=0.25.0",
    "class-registry>=2.1.2",
    "gcsfs>=2025.9.0",
    "hdf5plugin>=6.0.0",
    "matplotlib>=3.10.7",
    "pandas>=2.2",
    "psutil>=7.1.1",
    "rasterio>=1.4.3",
    "shapely>=2",
    "torchvision>=0.22.1",
    "universal-pathlib>=0.2.5",
    "wandb>=0.19.0",
]
beaker = [
    "beaker-py==1.34.1",
    "docker>=5.0,<8.0",
    "packaging>=24.2",
    "pydantic>=1.8.2,<3.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=12.3,<14.0",
    "google-cloud-compute",
]
dataset-creation = [
    "class-registry>=2.1.2",
    "einops==0.7.0",
    "gcsfs>=2025.9.0",
    "geobench>=1.0.1",
    "interrogate>=1.7.0",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "pandera>=0.26.1",
    "pytest>=8.4.2",
    "rasterio>=1.4.3",
    "rioxarray>=0.19.0",
    "rslearn==0.0.11",
    "torch>=2.7,<2.8",
    "universal-pathlib>=0.2.5",
    "wandb>=0.19.0",
]
dev = [
    "pip>=25.2",
    "gitpython>=3.0,<4.0",
    "interrogate>=1.7.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.1",
]
eval = [
    { include-group = "training" },
    "albumentations>=1.4.10",
    "claymodel",
    "geobench",
    "geopandas>=1.1.1",
    "pytorch-lightning==2.5.5",
    "pyyaml>=6.0",
    "rioxarray>=0.19.0",
    "rslearn==0.0.11",
    "satlaspretrain-models>=0.3.1",
    "scikit-learn>=1.7.2",
    "terratorch==1.1",
    "timm==1.0.20",
    "torchmetrics==1.7.1",
    "xarray>=2025.10.1",
]
flash_attn = ["flash-attn>=2.4.0"]

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }
